requirements.txt
matplotlib==3.10.0
netCDF4==1.7.2
numpy==2.1.3
PyYAML==6.0.2
streamhist==0.1.0
torch==2.5.1
wandb==0.23.1
welford==0.2.5
